Hormonize packages for WS module
* Improve `ws.adoc` to avoid `foo`, `bar` etc.
