+++ spring-integration-ws/src/main/java/org/springframework/integration/ws/inbound/package-info.java
@@ -1,0 +1,5 @@
+/**
+ * Inbound Web Service components.
+ */
+@org.jspecify.annotations.NullMarked
+package org.springframework.integration.ws.inbound;
